backend/requirements.txt
fastapi
requests
uvicorn
neo4j
pydantic
pydantic-settings
python-dotenv
numpy
yfinance
pandas
lxml
httpx
anyio
pytest
Jinja2
pytest-asyncio
langchain
langchain-core
langchain-google-genai
langchain-google-vertexai
langchain-community
langchain_openai
langgraph
langfuse
cloudscraper
beautifulsoup4
google-cloud-aiplatform
google-auth
scikit-learn
PyJWT[crypto]